requirements.txt
pandas==2.0.1
aiohttp==3.8.4
requests==2.30.0
pyarrow==14.0.1
python-dotenv